Fix player name support
The named-player support was broken (against current RaspberryJuice), because it
caused player.foo() to have the player name, or "None", as their first
parameter. This patch fixes this: if "name" is given to Minecraft.create(), then
the player.foo() commands will actually invoke the "entity" API package to
address the correct player. Otherwise (i.e., by default), the "player" API
package is used and the player name is stripped.